+++ src/main/java/com/wire/kalium/IEventProcessor.kt
@@ -1,8 +1,8 @@
 package com.wire.kalium
 
-import com.wire.kalium.backend.models.Payload
+import com.wire.kalium.models.backend.Payload
 import java.util.*
 
 interface IEventProcessor {
-    fun processEvent(eventId: UUID, payload: Payload, client: IWireClient)
+    fun processEvent(eventId: UUID, payload: Payload, wireClient: IWireClient)
 }
