+++ src/main/java/com/wire/kalium/IWireAPI.kt
@@ -1,8 +1,8 @@
 package com.wire.kalium
 
-import com.wire.kalium.backend.models.Conversation
-import com.wire.kalium.backend.models.User
 import com.wire.kalium.exceptions.HttpException
+import com.wire.kalium.models.backend.Conversation
+import com.wire.kalium.models.backend.User
 import com.wire.kalium.models.inbound.AssetKey
 import com.wire.kalium.models.outbound.Asset
 import com.wire.kalium.models.outbound.otr.*
