+++ src/main/java/com/wire/kalium/IWireClient.kt
@@ -18,9 +18,9 @@
 package com.wire.kalium
 
 import com.wire.bots.cryptobox.CryptoException
-import com.wire.kalium.backend.models.Conversation
-import com.wire.kalium.backend.models.User
 import com.wire.kalium.exceptions.HttpException
+import com.wire.kalium.models.backend.Conversation
+import com.wire.kalium.models.backend.User
 import com.wire.kalium.models.inbound.AssetKey
 import com.wire.kalium.models.outbound.Asset
 import com.wire.kalium.models.outbound.GenericMessageIdentifiable
