+++ src/main/java/com/wire/kalium/models/backend/Event.kt
@@ -1,6 +1,6 @@
 //
 // Wire
-// Copyright (C) 2020 Wire Swiss GmbH
+// Copyright (C) 2016 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,14 +15,15 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
-package com.wire.kalium.backend.models
+package com.wire.kalium.models.backend
 
+import com.wire.kalium.tools.UUIDSerializer
 import kotlinx.serialization.Serializable
+import java.util.*
 
 @Serializable
-data class AssetKey(
-        val type: String,
-        val key: String,
-        val size: String
-)
-
+class Event {
+    var id: @Serializable(with = UUIDSerializer::class) UUID? = null
+    var payload: Array<Payload>? = null
+    var transient = false
+}
