+++ src/main/java/com/wire/kalium/models/backend/Service.kt
@@ -1,6 +1,6 @@
 //
 // Wire
-// Copyright (C) 2016 Wire Swiss GmbH
+// Copyright (C) 2020 Wire Swiss GmbH
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,15 +15,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
-package com.wire.kalium.backend.models
+package com.wire.kalium.models.backend
 
 import com.wire.kalium.tools.UUIDSerializer
 import kotlinx.serialization.Serializable
 import java.util.*
 
 @Serializable
-class Event {
-    var id: @Serializable(with = UUIDSerializer::class) UUID? = null
-    var payload: Array<Payload>? = null
-    var transient = false
-}
+data class Service(
+        @Serializable(with = UUIDSerializer::class)  var id: UUID,
+        @Serializable(with = UUIDSerializer::class)  var provider: UUID
+)
