+++ src/main/java/com/wire/kalium/models/system/SystemMessage.kt
@@ -1,0 +1,17 @@
+package com.wire.kalium.models.system
+
+import com.wire.kalium.models.backend.Conversation
+import com.wire.kalium.tools.UUIDSerializer
+import kotlinx.serialization.Serializable
+import java.util.*
+
+@Serializable
+data class SystemMessage(
+        @Serializable(with = UUIDSerializer::class) val id: UUID,
+        val type: String,
+        val time: String,
+        @Serializable(with = UUIDSerializer::class) val from: UUID,
+        val conversation: Conversation,
+        @Serializable(with = UUIDSerializer::class) val convId: UUID,
+        val userIds: List<@Serializable(with = UUIDSerializer::class) UUID>
+)
